Added Session Hash and error logging and removed hard coded key
Added Session Has field to both Legiscan Session and Model Session
to allow for change detection

Added Error Logging to LegiscanHttpRequest

Removed hard coded key which was added to allow for testing.  Now when
the key is requested, we check to see if the Spring Context has loaded,
and if not, manually pull it from the properties file.